refactor(subgraph): move intmax utility functions to separate file
